+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [0.8.0] - 2018-07-
 
-### Changed
+### Removed
 
 - Removed mypayex logo option from topbar. Logo is now defaulted to standard payex logo, class `.logo-default` is no longer needed.
+- Removed `.disabled` class styling for pagination component arrows. `span` elements will now act as disabled arrows.
 
 ## [0.7.1] - 2018-07-03