+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Changed `.background-color` utility class to `.bg-color`.
 - Changed breadcrumb component to disable span elements by default, provided they are not within last-child li element.
 - Reworked styling of tabs component.
+- Reworked alert component.
 
 ### Fixed